+++ examples/fqdnGateway/Pulumi.yaml
@@ -11,28 +11,14 @@ resources:
     type: pulumi:providers:grid
     properties:
       mnemonic:
-      network: qa
-
-  network:
-    type: grid:internal:Network
-    options:
-      provider: ${provider}
-    properties:
-      name: test
-      description: test network
-      nodes:
-        - 3
-      ip_range: 10.1.0.0/16
 
   fqdnGateway:
     type: grid:internal:FqdnGateway
     options:
       provider: ${provider}
-      dependsOn:
-        - ${network}
     properties:
       name: gateway
-      node_id: 2
-      fqdn: domain
+      node_id: 3
+      fqdn: test.gent01.qa.grid.tf
       backends:
-        - https://10.1.0.2:8081
+        - http://69.164.223.208
