+++ internal/fqdn_gateway_resource.go
@@ -1,12 +1,8 @@
 package provider
 
 import (
-	"fmt"
-	"strconv"
-
 	p "github.com/pulumi/pulumi-go-provider"
 	"github.com/pulumi/pulumi-go-provider/infer"
-	"github.com/threefoldtech/tfgrid-sdk-go/grid-client/workloads"
 	"github.com/threefoldtech/zos/pkg/gridtypes/zos"
 )
 
@@ -31,88 +27,6 @@ type FqdnGatewayState struct {
 
 	ContractID       int64            `pulumi:"contract_id"`
 	NodeDeploymentID map[string]int64 `pulumi:"node_deployment_id"`
-}
-
-func parseToFqdnGatewayState(fqdnGateway workloads.GatewayFQDNProxy) FqdnGatewayState {
-
-	stateArgs := FqdnGatewayArgs{
-		NodeID:         int32(fqdnGateway.NodeID),
-		Name:           fqdnGateway.Name,
-		Description:    fqdnGateway.Description,
-		SolutionType:   fqdnGateway.SolutionType,
-		NetworkName:    fqdnGateway.Network,
-		TLSPassthrough: fqdnGateway.TLSPassthrough,
-		FQDN:           fqdnGateway.FQDN,
-		Backends:       fqdnGateway.Backends,
-	}
-
-	nodeDeploymentID := make(map[string]int64)
-	for key, value := range fqdnGateway.NodeDeploymentID {
-		nodeDeploymentID[fmt.Sprint(key)] = int64(value)
-	}
-
-	state := FqdnGatewayState{
-		FqdnGatewayArgs:  stateArgs,
-		ContractID:       int64(fqdnGateway.ContractID),
-		NodeDeploymentID: nodeDeploymentID,
-	}
-
-	return state
-
-}
-
-func parseToWorkloadFqdnGateway(fqdnGatewayArgs FqdnGatewayArgs) workloads.GatewayFQDNProxy {
-
-	return workloads.GatewayFQDNProxy{
-
-		NodeID:         uint32(fqdnGatewayArgs.NodeID),
-		Name:           fqdnGatewayArgs.Name,
-		SolutionType:   fqdnGatewayArgs.SolutionType,
-		Network:        fqdnGatewayArgs.NetworkName,
-		FQDN:           fqdnGatewayArgs.FQDN,
-		Backends:       fqdnGatewayArgs.Backends,
-		TLSPassthrough: fqdnGatewayArgs.TLSPassthrough,
-		Description:    fqdnGatewayArgs.Description,
-	}
-
-}
-
-func parseToFqdnGatewayComputed(fqdnGatewayState FqdnGatewayState) workloads.GatewayFQDNProxy {
-
-	nodeDeploymentID := make(map[uint32]uint64)
-
-	for key, value := range fqdnGatewayState.NodeDeploymentID {
-		k, err := strconv.ParseUint(key, 10, 32)
-		if err != nil {
-			continue
-		}
-		nodeDeploymentID[uint32(k)] = uint64(value)
-	}
-
-	return workloads.GatewayFQDNProxy{
-		NodeDeploymentID: nodeDeploymentID,
-		ContractID:       uint64(fqdnGatewayState.ContractID),
-	}
-
-}
-
-func updateFqdnGatewayFromState(fqdnGateway *workloads.GatewayFQDNProxy, state FqdnGatewayState) error {
-
-	nodeDeploymentID := make(map[uint32]uint64)
-
-	for key, value := range state.NodeDeploymentID {
-		k, err := strconv.ParseUint(key, 10, 32)
-		if err != nil {
-			continue
-		}
-		nodeDeploymentID[uint32(k)] = uint64(value)
-	}
-
-	fqdnGateway.ContractID = uint64(state.ContractID)
-	fqdnGateway.NodeDeploymentID = nodeDeploymentID
-
-	return nil
-
 }
 
 // Create creates a fqdn gateway
